refactor: move contact-preview-actions.tsx components to views/contacts and rearrange contact-list hover actions
- it is time to stop using "legacy" folder
- we can start reusing the same actions defined in the preview
